==> IMC/04_02_cluster_Tcells_mesmer.Rmd <==
---
title: "R Notebook - Cluster T cells"
output:
  html_document:
    df_print: paged
---

```{r, import libraries, message=F, echo=F, warning=F}
library(igraph)
library(SingleCellExperiment)
library(S4Vectors)
library(stringr)
library(DT)
library(dplyr)
library(tidyr)
library(mclust)
library(ggplot2)
library(RColorBrewer)
library(scater)
library(Rphenoannoy)
set.seed(101100)
```


```{r, Set wd and load data}
#set working directory 
wd <-dirname(getwd())

data_folder <-(file.path(wd,"sce_objects","immune","Tcell"))

#RAW
#all.tcell <- readRDS(file=file.path(data_folder, "tcells_RAW.rds"))

#workingfile
all.tcell <- readRDS(file=file.path(data_folder, "tcells_workingfile.rds"))
#saveRDS(all.tcell, file=file.path(data_folder, "all_tcell_workingfile.rds"))
```


```{r, Define immune markers, echo=F, warning=F, message=FALSE}
all.marker <-rownames(all.tcell)

bad.marker <- c("Iridium191","Iridium193","CD11b","Histone H3") 
good.marker <- all.marker[!all.marker %in% bad.marker]
immune.marker <-c( "CD11b_1","CD20","HLA-DR","CD68","IDO", "CD3","CD11c", "MMP9","FOXP3","CXCL13","CD8","CD4","Ki-67","CD45","Myeloperoxidase MPO")

tcell.marker <- c("IDO", "CD3","FOXP3","CXCL13","CD8","CD4","Ki-67")
#tbcell.marker <- c("IDO", "CD3","FOXP3","CXCL13","CD8","CD4","Ki-67","CD20")

print(immune.marker)
print(good.marker)
```

```{r, calculate umap all tcells, warning=F, message=F, echo=F, eval=FALSE}
p <-c(10,50,100)
#p=50
for(i in p){
all.tcell <- runUMAP(all.tcell,
                 exprs_values = "c_counts_asinh",
                 name = paste0("UMAP_p", i),
                 #use_dimred="PCA_20",
                 n_neighbors = i)
saveRDS(all.tcell, file=file.path(data_folder,"tcells_workingfile.rds"))

}
saveRDS(all.tcell, file=file.path(data_folder,"tcells_workingfile.rds"))


for(i in p){
all.tcell <- runTSNE(all.tcell.sub,
                 exprs_values = "c_counts_asinh",
                 name = paste0("tSNE_p", i),
                 #use_dimred="PCA_20",
                 perplexity = i)
saveRDS(all.tcell, file=file.path(data_folder,"Tcell","tcells_workingfile.rds"))
}
```


**UMAP with good markers**
```{r,plot umap all t cells immune marker, fig.width=12, fig.height=8, echo=F, message=FALSE,warning=FALSE}
dat <-as.data.frame(reducedDims(all.tcell)$`UMAP_p10`)
dat$cell <- rownames(dat)
dat.counts <-as.data.frame(t((assay(all.tcell,"c_counts_asinh_scaled"))))
dat.counts$cell <- rownames(dat.counts)
dat.all <-merge(dat, dat.counts, by.x="cell", by.y="cell")

dat.all.long <-pivot_longer(dat.all, cols = colnames(dat.all)[colnames(dat.all) %in% tcell.marker], names_to = "target", values_to = "counts")


p <-dat.all.long %>%
  ggplot(aes(x=V1, y=V2, color=counts))+
  facet_wrap(~target, scales = "free", ncol = 4)+
  geom_point(alpha=0.5, size=0.2)+
  scale_color_gradientn(colours=rev(brewer.pal(11, 'Spectral')), name='Counts')+
  ggtitle('')+
  theme(strip.background = element_blank(),
        axis.line=element_blank(),
        axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        axis.title.x=element_blank(),
        axis.title.y=element_blank(),
        panel.background=element_blank(),
        panel.border=element_blank(),
        panel.grid.major=element_blank(),
        panel.grid.minor=element_blank(),
        plot.background=element_blank(),
         strip.text = element_text(size=8)) 

##ggsave(filename=file.path(plot_folder, paste("sub_fibro_fibro-Marker_UMAP_p50_NEW.png",sep="")), plot=p, width=16, height=10)
plot(p)
```

**tsne with tb markers**
```{r,plot tsne t cell marker, fig.width=12, fig.height=8, echo=F, message=FALSE,warning=FALSE,eval=F}
dat <-as.data.frame(reducedDims(all.tcell.sub)$`tSNE_p50`)
dat$cell <- rownames(dat)
dat.counts <-as.data.frame(t((assay(all.tcell.sub,"c_counts_asinh_scaled"))))
dat.counts$cell <- rownames(dat.counts)
dat.all <-merge(dat, dat.counts, by.x="cell", by.y="cell")

dat.all.long <-pivot_longer(dat.all, cols = colnames(dat.all)[colnames(dat.all) %in% good.marker], names_to = "target", values_to = "counts")


p <-dat.all.long %>%
  ggplot(aes(x=V1, y=V2, color=counts))+
  facet_wrap(~target, scales = "free", ncol = 6)+
  geom_point(alpha=0.5, size=0.2)+
  scale_color_gradientn(colours=rev(brewer.pal(11, 'Spectral')), name='Counts')+
  ggtitle('')+
  theme(strip.background = element_blank(),
        axis.line=element_blank(),
        axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        axis.title.x=element_blank(),
        axis.title.y=element_blank(),
        panel.background=element_blank(),
        panel.border=element_blank(),
        panel.grid.major=element_blank(),
        panel.grid.minor=element_blank(),
        plot.background=element_blank(),
         strip.text = element_text(size=8)) 

##ggsave(filename=file.path(plot_folder, paste("sub_fibro_fibro-Marker_UMAP_p50_NEW.png",sep="")), plot=p, width=16, height=10)
plot(p)
```


```{r Clustering all tcell using Rphenoannoy, fig.width=25, fig.height=12 , message=F, echo=F, warning=F, eval=F}
hmcol<-rev(brewer.pal(11,"RdBu"))
rdylbu <-rev(brewer.pal(11,"RdYlBu"))

#marker.clustering <- good.marker[!good.marker %in% "Pan Cytokeratin + Keratin Epithelial"]
#cellnb <- 250000
#subsample <-  colnames(sce)[sample(length(colnames(sce)), cellnb)]
#sce_sub <- sce[, colnames(sce) %in% subsample]

#set ks for clustering
#i <- 100
k <- c(10,15,20,25,30)
k <- c(11:14, 16:19, 21:24, 26:29, 31:35)
#i<-10
for (i in k) {
    all.tcell$RPmembership <- factor(Rphenoannoy(data = t(assay(all.tcell[rownames(all.tcell) %in% tcell.marker,],"c_counts_asinh")), k = i)[[2]]$membership)
    cluster <- paste0("rp_tcell_k",i)
  colnames(colData(all.tcell))[which(names(colData(all.tcell)) == "RPmembership")] <- paste0(cluster)
  agg_sce <-aggregateAcrossCells(all.tcell, ids=all.tcell[[cluster]], average=TRUE, use_exprs_values="c_counts_asinh_scaled" )

 
    scater::plotHeatmap(agg_sce,
              features = tcell.marker,
              #features = immune.marker,
              #features = marker.clustering,
              exprs_values = "c_counts_asinh_scaled",
              symmetric = FALSE,
              zlim=c(0,1),
              color=rdylbu,
              sortrowss=FALSE,
              show_colnames=TRUE,
              color_columns_by = c(cluster),
              main=paste0("Heatmap tumour cells, ",cluster))
    
  #save sce clustering
#saveRDS(all.tcell, file=file.path(data_folder,"tcells_workingfile.rds"))
}
#all.tcell <-readRDS( file=file.path(data_folder, "all_tcell_workingfile.rds"))
saveRDS(all.tcell, file=file.path(data_folder,"tcells_workingfile.rds"))

```

```{r, assign cluster levels T cell final,fig.width=12, fig.height=10, message=F, echo=F, warning=F}
hmcol<-rev(brewer.pal(11,"RdBu"))
rdylbu <-rev(brewer.pal(11,"RdYlBu"))


is.CD8 <- c(7,5,10,13)
is.CD8.IDO <- c(18)
is.CD8.ki67 <- c(12)
is.CD8.Treg <- 4
is.CD8.CXCL13 <- c(15)

is.CD4.Treg <-c(2,1)
is.CD4 <-c(22,9,6,3, 20,21,8)
is.CD4.ki67 <-c(11)
is.CD4.IDO <-c(17)
is.CD4.CD8 <- 19
is.CD4.CXCL13 <- 16

is.other <- c(14)

all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD8] <-"CD8"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD8.Treg] <-"CD8_Treg"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD8.IDO] <-"IDO_CD8"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD8.ki67] <-"ki67_CD8"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD8.CXCL13] <-"CD8_CXCL13"

all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD4.Treg] <-"CD4_Treg"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD4] <-"CD4"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD4.ki67] <-"ki67_CD4"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD4.IDO] <-"IDO_CD4"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD4.CXCL13] <-"CD4_CXCL13"

all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.CD4.CD8] <-"T cell"
all.tcell$TcellType[all.tcell$rp_tcell_k35%in% is.other] <-"other"

agg_sce <-aggregateAcrossCells(all.tcell, ids=all.tcell$rp_tcell_k35, average=TRUE, use_exprs_values="c_counts_asinh_scaled" )

 
    scater::plotHeatmap(agg_sce,
              features = tcell.marker,
              #features = immune.marker,
              #features = marker.clustering,
              exprs_values = "c_counts_asinh_scaled",
              symmetric = FALSE,
              zlim=c(0,1),
              color=rdylbu,
              sortrowss=FALSE,
              show_colnames=TRUE,
              color_columns_by = c("TcellType"))
    
    
agg_sce <-aggregateAcrossCells(all.tcell, ids=all.tcell$TcellType, average=TRUE, use_exprs_values="c_counts_asinh_scaled" )

 
    scater::plotHeatmap(agg_sce,
              features = tcell.marker,
              #features = immune.marker,
              #features = marker.clustering,
              exprs_values = "c_counts_asinh_scaled",
              symmetric = FALSE,
              zlim=c(0,1),
              color=rdylbu,
              sortrowss=FALSE,
              show_colnames=TRUE,
              color_columns_by = c("TcellType"))
    
    
#saveRDS(all.tcell.immune, file=file.path(data_folder, "tcell-final-immune.rds"))

#saveRDS(all.tcell, file=file.path(data_folder, "FINAL_Tcells-minusImmune_RAW.rds"))
#saveRDS(all.tcell, file=file.path(data_folder, "FINAL_Tcells-minusImmune_workingfile.rds"))


#T cell category
all.tcell$TcellCategory[all.tcell$rp_tcell_k35 %in% is.CD8|
                      all.tcell$rp_tcell_k35 %in% is.CD8.Treg|
                      all.tcell$rp_tcell_k35 %in% is.CD8.IDO|
                      all.tcell$rp_tcell_k35 %in% is.CD8.ki67|
                      all.tcell$rp_tcell_k35 %in% is.CD8.CXCL13] <-"CD8"


all.tcell$TcellCategory[all.tcell$rp_tcell_k35%in% is.CD4.Treg|
                      all.tcell$rp_tcell_k35%in% is.CD4|
                      all.tcell$rp_tcell_k35%in% is.CD4.ki67|
                      all.tcell$rp_tcell_k35%in% is.CD4.IDO|
                        all.tcell$rp_tcell_k35 %in% is.CD4.CXCL13
                      ] <-"CD4"

all.tcell$TcellCategory[all.tcell$rp_tcell_k35%in% is.other] <-"other"
all.tcell$TcellCategory[all.tcell$rp_tcell_k35 %in% is.CD4.CD8] <- "T cell"

table(all.tcell$TcellCategory)
table(all.tcell$TcellType)

agg_sce <-aggregateAcrossCells(all.tcell, ids=all.tcell$TcellType, average=TRUE, use_exprs_values="c_counts_asinh_scaled" )

 
    scater::plotHeatmap(agg_sce,
              features = tcell.marker,
              #features = immune.marker,
              #features = marker.clustering,
              exprs_values = "c_counts_asinh_scaled",
              symmetric = FALSE,
              zlim=c(0,1),
              color=rdylbu,
              sortrowss=FALSE,
              show_colnames=TRUE,
              color_columns_by = c("TcellType","TcellCategory"))
```
```{r, plot umap with t cell category and types minus immune, message=F, warning=F, echo=F}
#Tell Category
cluster<- "rp_tcell_k35"
p <-plotReducedDim(all.tcell, "UMAP_p10", colour_by=paste(cluster), text_by = paste(cluster),text_colour ="red", point_size=0.5,)+labs(x="UMAP-1", y="UMAP-2", title="UMAP") 
plot(p)

#Tell Category
cluster<- "TcellCategory"
p <-plotReducedDim(all.tcell, "UMAP_p10", colour_by=paste(cluster), text_by = paste(cluster),text_colour ="red", point_size=0.5,)+labs(x="UMAP-1", y="UMAP-2", title="UMAP") 
plot(p)

#Tell Type
cluster<- "TcellType"
p <-plotReducedDim(all.tcell, "UMAP_p10", colour_by=paste(cluster), text_by = paste(cluster),text_colour ="red", point_size=0.5,)+labs(x="UMAP-1", y="UMAP-2", title="UMAP") 
plot(p)
```

```{r}
saveRDS(all.tcell, file=file.path(data_folder, "all_tcell_workingfile.rds"))
```
